Answers for standard_library_types
iterators4.rs was particularly instructive.

I started out by searching for the initial C recursion and for loop
versions and tried to use a range iteration but couldn't figure the
syntax. So I created the first 7 answers and then used hint.

I then seached for: how to use iterator for side effects only
 https://www.google.com/search?q=how+to+use+iterator+for+side+effects+only+in+rust

And saw this:
 https://github.com/rust-lang/rust/issues/44546#issuecomment-344192206
And:
 https://github.com/rust-lang/rust/issues/44546#issuecomment-381678863

And with that I implemented Answer 8

But still wasn't happy as I felt there must be a better way where I didn't
need to throw away things. So I looking at the iterator documentation:
 https://doc.rust-lang.org/std/iter/trait.Iterator.html

I saw "fold" and low and behold that was exactly what I needed! Then I
saw "fold_first" and it was "simpler" as the first value in the range
is the initial value for the accumulator so fold_first only has one
parameters not two. But fold_first returns an Option which means I need
to return the unwrap(). Anyway both Answers 9 and 10 are simpler then
any other solution!

I do wonder which is fastest.